Refactor out offset module, remove  cipher tests and Cipher class and move to Enigma test and class files
